improved the Note entity

+++ server/src/main/java/server/database/NoteRepository.java
@@ -3,5 +3,5 @@ package server.database;
 import commons.Note;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NoteRepository extends JpaRepository<Note, Integer> {
+public interface NoteRepository extends JpaRepository<Note, Long> {
 }
